refactor: change folder name from 'contants' to 'constants'
