+++ src/constants/slider.ts
